refactor(recovery): extract advanced conditional-format handlers
